+++ platform/core/src/__mocks__/log.js
@@ -1,4 +1,5 @@
 export default {
   warn: jest.fn(),
   error: jest.fn(),
+  info: jest.fn(),
 };