+++ platform/ui/src/components/tabComponents/index.js
@@ -1,0 +1,2 @@
+export { TabComponents } from './TabComponents';
+export { TabFooter } from './TabFooter';