+++ platform/viewer/src/components/Header/Header.js
@@ -1,10 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import { Link, withRouter } from 'react-router-dom';
 import { withTranslation } from 'react-i18next';
+
 import PropTypes from 'prop-types';
 
-import ConnectedUserPreferencesForm from '../../connectedComponents/ConnectedUserPreferencesForm';
 import { Dropdown, AboutContent, withModal } from '@ohif/ui';
+
+import { UserPreferences } from './../UserPreferences';
 import OHIFLogo from '../OHIFLogo/OHIFLogo.js';
 import './Header.css';
 
@@ -42,7 +44,7 @@ function Header(props) {
         },
         onClick: () =>
           show({
-            content: ConnectedUserPreferencesForm,
+            content: UserPreferences,
             title: t('User Preferences'),
           }),
       },